factor out the UI Builder into its own module -- just get it working for now.
